app/requirements.txt
pandas
matplotlib
matplotlib-inline
pydantic
pydantic-settings
python-dotenv
plotly
dash
dash-bootstrap-components
dash-ag-grid
influxdb-client
Flask
Werkzeug
gunicorn

# for development
ipython
ipdb
